[add] admin touristテスト
tourist関連ページテストの実装
